chapter_004/src/test/java/ru/job4j/srp/ReportPersonalTest.java
package ru.job4j.srp;

import static org.junit.Assert.assertThat;
import static org.hamcrest.Matchers.is;
import org.junit.Test;
import java.util.Calendar;

public class ReportPersonalTest {

    @Test
    public void generate() {
        MemStore store = new MemStore();
        Calendar now = Calendar.getInstance();
        Employee worker = new Employee("Ivan", now, now, 100);
        Employee worker1 = new Employee("MiKl", now, now, 300);
        Employee worker2 = new Employee("Sam", now, now, 400);
        Employee worker3 = new Employee("Ann", now, now, 250);
        store.add(worker);
        store.add(worker1);
        store.add(worker2);
        store.add(worker3);
        Report personal = new ReportPersonal(store);
        StringBuilder expect = new StringBuilder()
                .append("Name; Salary")
                .append(System.lineSeparator())
                .append(worker.getName()).append(";")
                .append(worker.getSalary()).append(";")
                .append(System.lineSeparator())
                .append(worker1.getName()).append(";")
                .append(worker1.getSalary()).append(";")
                .append(System.lineSeparator())
                .append(worker2.getName()).append(";")
                .append(worker2.getSalary()).append(";")
                .append(System.lineSeparator())
                .append(worker3.getName()).append(";")
                .append(worker3.getSalary()).append(";")
                .append(System.lineSeparator());
        assertThat(personal.generate(em -> true), is(expect.toString()));
    }
}